requirements.txt
django
django-heroku
python-dotenv
gunicorn
django-heroku
djangorestframework
pillow
whitenoise
daphne
django-sslserver
psycopg2-binary
dj-database-url
django-storages
boto3
selenium
chromedriver-py
arrow
channels
channels-redis
django-cors-headers
django-rest-resetpassword
google-oauth2-tool
google-auth
social-auth-app-django